Database Fixes
Se arreglaron algunos errores del diagrama relacional y ER, se cambio el codigo de la base de datos respectivamente, asi como empezar el desarrollo por modulos de nuestras clases, iniciando en la clase que maneja la base de datos para implementar los procedimientos de insercion y borrado, y metodos de actualizacion

+++ AplicacionWeb/Database/GreenZenithProcedures.sql
@@ -10,19 +10,14 @@ DELIMITER ;
 
 DELIMITER $$
 DROP PROCEDURE IF EXISTS insertUser $$
-CREATE PROCEDURE insertUser (IN Username NVARCHAR(50), IN Email NVARCHAR(100), IN Age INT, IN Location NVARCHAR(50), IN PasswordUser NVARCHAR(30))
+CREATE PROCEDURE insertUser (IN Username NVARCHAR(50), IN Email NVARCHAR(100), IN Age INT, IN Location NVARCHAR(50), IN PasswordUser NVARCHAR(30), IN AdministratorAccess BOOLEAN)
 BEGIN
-INSERT INTO PUser VALUES (NULL, Username, Email, Age, Location, PasswordUser);
+INSERT INTO PUser VALUES (NULL, Username, Email, Age, Location, PasswordUser, AdministratorAccess);
 END $$
 DROP PROCEDURE IF EXISTS insertUserPicture $$
 CREATE PROCEDURE insertUserPicture (IN UserID INT, IN Picture MEDIUMBLOB)
 BEGIN
 INSERT INTO UserPicture VALUES (UserID, Picture);
-END $$
-DROP PROCEDURE IF EXISTS insertIntoAdministratorAccess $$
-CREATE PROCEDURE insertIntoAdministratorAccess (IN UserID INT)
-BEGIN
-INSERT INTO AdministratorAccess VALUES (UserID);
 END $$
 
 DROP PROCEDURE IF EXISTS insertPlant $$
@@ -52,15 +47,10 @@ BEGIN
 INSERT INTO ProductPicture VALUES (ProductID, Picture);
 END $$
 
-DROP PROCEDURE IF EXISTS insertMessageSender $$
-CREATE PROCEDURE insertMessageSender (IN Content TEXT, SentDateTime DATETIME, SenderID INT)
+DROP PROCEDURE IF EXISTS insertMessage $$
+CREATE PROCEDURE insertMessages (IN Content TEXT, SentDateTime DATETIME, SenderID INT, Reciever INT)
 BEGIN
-INSERT INTO MessageSender VALUES (NULL, Content, SentDateTime, SenderID);
-END $$
-DROP PROCEDURE IF EXISTS insertMessageReciever $$
-CREATE PROCEDURE insertMessageReciever (IN RecieverID INT, IN SenderID INT)
-BEGIN
-INSERT INTO MessageReciever VALUES (RecieverID, SenderID);
+INSERT INTO Messages VALUES (NULL, Content, SentDateTime, SenderID, RecieverID);
 END $$
 
 DROP PROCEDURE IF EXISTS insertIntoCart $$
@@ -78,11 +68,6 @@ DROP PROCEDURE IF EXISTS deleteUserPicture $$
 CREATE PROCEDURE deleteUserPicture (IN UserID INT)
 BEGIN
 DELETE FROM PUserPicture WHERE PUserPicture.UserID = UserID;
-END $$
-DROP PROCEDURE IF EXISTS deleteAdministratorAccess $$
-CREATE PROCEDURE deleteAdministratorAccess (IN UserID INT)
-BEGIN
-DELETE FROM AdministratorAccess WHERE AdministratorAccess.UserID = UserID;
 END $$
 
 DROP PROCEDURE IF EXISTS deletePlant $$
@@ -112,12 +97,11 @@ BEGIN
 DELETE FROM ProductPicture WHERE ProductPicture.ProductID = ProductID;
 END $$
 
-DROP PROCEDURE IF EXISTS deleteMessageSender $$
-CREATE PROCEDURE deleteMessageSender (IN MessageID INT)
+DROP PROCEDURE IF EXISTS deleteMessage $$
+CREATE PROCEDURE deleteMessages (IN MessageID INT)
 BEGIN
-DELETE FROM MessageSender WHERE ID = MessageID;
+DELETE FROM Messages WHERE ID = MessageID;
 END $$
-DROP PROCEDURE IF EXISTS deleteMessageReciever $$
 
 
 DROP PROCEDURE IF EXISTS deleteCart $$
@@ -126,8 +110,8 @@ BEGIN
 DELETE FROM Cart WHERE Cart.UserID = UserID;
 END $$
 DROP PROCEDURE IF EXISTS deleteCartProduct $$
-CREATE PROCEDURE deleteCartProduct (IN ProductID INT)
+CREATE PROCEDURE deleteCartProduct (IN UserID INT, IN ProductID INT)
 BEGIN
-DELETE FROM Cart WHERE Cart.ProductID = ProductID;
+DELETE FROM Cart WHERE Cart.ProductID = ProductID AND Cart.UserID = UserID;
 END $$
 DELIMITER ;
